+++ engine/flink/management/src/main/scala/pl/touk/nussknacker/engine/management/FlinkRestManager.scala
@@ -86,7 +86,9 @@ class FlinkRestManager(config: FlinkConfig, modelData: ModelData, mainClassName:
   //NOTE: Flink <1.10 compatibility - protected to make it easier to work with Flink 1.9, JobStatus changed package, so we use String in case class
   protected def mapJobStatus(overview: JobOverview): StateStatus = {
     toJobStatus(overview) match {
-      case JobStatus.RUNNING => FlinkStateStatus.Running
+      case JobStatus.RUNNING if overview.tasks.running == overview.tasks.total => FlinkStateStatus.Running
+      // Flink return running status even if some tasks are scheduled or initializing
+      case JobStatus.RUNNING => FlinkStateStatus.DuringDeploy
       case JobStatus.FINISHED => FlinkStateStatus.Finished
       case JobStatus.RESTARTING => FlinkStateStatus.Restarting
       case JobStatus.CANCELED => FlinkStateStatus.Canceled